Rename gamepad and node-gamepad types
BREAKING CHANGE: `gamepad` has been renamed to `w3c-gamepad`.

BREAKING CHANGE: `node-gamepad` has been renamed to `gamepad`.
